Feedback code review part3
