+++ 08.ls_object/lib/ls/horizontal_formatter.rb
@@ -3,26 +3,21 @@
 require_relative 'formatter'
 
 class LsHorizontalFormatter < LsFormatter
-  def initialize(file_paths, width)
+  def self.run(file_paths, width)
     @file_paths = file_paths
     @file_names = []
     @width = width
     build_file_name
-  end
-
-  def run
     format_file_name
   end
 
-  private
-
-  def build_file_name
+  def self.build_file_name
     @file_names = @file_paths.map do |file_path|
       File.basename(file_path)
     end
   end
 
-  def format_file_name
+  def self.format_file_name
     max_file_name_length = find_max_length(@file_names)
     col_count = @width / (max_file_name_length + 1)
     row_count = col_count.zero? ? @file_names.count : (@file_names.count.to_f / col_count).ceil
@@ -32,17 +27,17 @@ class LsHorizontalFormatter < LsFormatter
     render_row_data(transposed_file_names)
   end
 
-  def align_file_name(max_file_name_length)
+  def self.align_file_name(max_file_name_length)
     @file_names.map do |file_name|
       file_name.to_s.ljust(max_file_name_length)
     end
   end
 
-  def segment_file_name(aligned_file_names, row_count)
+  def self.segment_file_name(aligned_file_names, row_count)
     aligned_file_names.each_slice(row_count).to_a
   end
 
-  def safe_transpose(segmented_file_name)
+  def self.safe_transpose(segmented_file_name)
     segmented_file_name[0].zip(*segmented_file_name[1..-1])
   end
 end
